requirements.txt
sqlalchemy
python-dotenv
apscheduler
requests
groq
pydantic
pydantic-settings
python-multipart
beautifulsoup4
resend
googlesearch-python
ratelimit
aiohttp